t4 float korvattu int

+++ Moduuli2/Tehtävä4.py
@@ -5,7 +5,7 @@ luku2 = int(input("Syötä toinen kokonaisluku: "))
 luku3 = int(input("Syötä kolmas kokonaisluku: "))
 summa = luku1 + luku2 + luku3
 tulo = luku1 * luku2 * luku3
-keskiarvo = int(summa / 3)
+keskiarvo = (summa / 3)
 print("Lukujen summa on: " + str(summa))
 print("Lukujen tulo on: " + str(tulo))
 print("Lukujen keskiarvo on: " + str(keskiarvo))
